requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
influxdb-client==1.25.0
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
Rx==3.2.0
six==1.16.0
toml==0.10.2
urllib3==1.26.8
